start to interactive subcomands (kinda)
Code may be a little bad but this is just a start as I don't really understand this system. But I need you ( @booboojp ) to finish this. But I need it so when you run the ysws create command it will start the interactive mode but when you run say ysws list it won't as that will list all projects (lmited by pagination).